# australia-chat-backend/api.yaml
swagger: "2.0"
info:
  title: Dating API
  description: This is Dating API Documentation
  version: one
schemes: [http, https]
paths:
  /api/v1/users/signUp/SendOTP:
    post:
      tags:
        - User Authentication
      summary: Sign Up send Otp on phone no
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            type: object
            required:
              - phone
            properties:
              phone:
                type: number
      responses:
        "200":
          description: Otp Is sended to phone id and Otp is valid for 10 min
        "409":
          description: This phone number already exists
        "500":
          description: Internal server error
...